backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx>=0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psutil==6.1.0
